:white_check_mark: :recycle: Extract UpdateDbTest
