+++ modules/cli/src/test/kotlin/datamaintain/cli/update/db/UpdateDbTest.kt
@@ -1,0 +1,399 @@
+package datamaintain.cli.update.db
+
+import datamaintain.cli.BaseCliTest
+import datamaintain.core.config.DatamaintainConfig
+import datamaintain.core.script.ScriptAction
+import datamaintain.core.script.Tag
+import datamaintain.core.script.TagMatcher
+import datamaintain.core.step.check.rules.implementations.SameScriptsAsExecutedCheck
+import datamaintain.db.driver.mongo.MongoDriverConfig
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.EnumSource
+import strikt.api.expectThat
+import strikt.assertions.*
+import java.nio.file.Paths
+import kotlin.reflect.KProperty1
+
+internal class UpdateDbTest : BaseCliTest() {
+    @Nested
+    inner class ConfigurationBuild {
+        @Nested
+        inner class GenericConfiguration {
+            @Test
+            fun `should build config with path`() {
+                // Given
+                val path = "myPath"
+
+                val argv = updateMongoDbMinimumArguments().plus(
+                    listOf(
+                        "--path", path
+                    )
+                )
+
+                // When
+                runApp(argv)
+
+                // Then
+                expectThat(configWrapper.datamaintainConfig!!.path).isEqualTo(Paths.get(path))
+            }
+
+            @Test
+            fun `should build config with identifier regex`() {
+                // Given
+                val identifierRegex = "myIdentifierRegex"
+
+                val argv = updateMongoDbMinimumArguments().plus(
+                    listOf(
+                        "--identifier-regex", identifierRegex
+                    )
+                )
+
+                // When
+                runApp(argv)
+
+                // Then
+                expectThat(configWrapper.datamaintainConfig!!.identifierRegex.pattern).isEqualTo(identifierRegex)
+            }
+
+            @Nested
+            inner class ExecutionMode {
+                @ParameterizedTest
+                @EnumSource(
+                    datamaintain.core.step.executor.ExecutionMode::class,
+                    names = ["FORCE_MARK_AS_EXECUTED"],
+                    mode = EnumSource.Mode.EXCLUDE
+                )
+                @DisplayName("Should build config with execution mode {0}")
+                fun `should build config with execution mode`(executionMode: datamaintain.core.step.executor.ExecutionMode) {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus(
+                        listOf(
+                            "--execution-mode", executionMode.name
+                        )
+                    )
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper.datamaintainConfig!!.executionMode).isEqualTo(executionMode)
+                }
+
+                @Test
+                fun `should build config with FORCE_MARK_AS_EXECUTED as execution mode`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus(
+                        listOf(
+                            "--execution-mode",
+                            datamaintain.core.step.executor.ExecutionMode.FORCE_MARK_AS_EXECUTED.name
+                        )
+                    )
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper.datamaintainConfig!!).and {
+                        get { executionMode }.isEqualTo(datamaintain.core.step.executor.ExecutionMode.NORMAL)
+                        get { defaultScriptAction }.isEqualTo(ScriptAction.MARK_AS_EXECUTED)
+                    }
+                }
+            }
+
+            @Nested
+            inner class Verbose {
+                @Test
+                fun `should build config with verbose set to true`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus("--verbose")
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper.datamaintainConfig!!.verbose).isTrue()
+                }
+
+                @Test
+                fun `should build config with verbose set to false`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments()
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper.datamaintainConfig!!.verbose).isFalse()
+                }
+            }
+
+            @Nested
+            inner class CreateTagsFromFolder {
+                @Test
+                fun `should build config with create tags from folder set to true`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus("--create-tags-from-folder")
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper.datamaintainConfig!!.doesCreateTagsFromFolder).isTrue()
+                }
+
+                @Test
+                fun `should build config with create tags from folder set to false`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments()
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper.datamaintainConfig!!.doesCreateTagsFromFolder).isFalse()
+                }
+            }
+
+            @Nested
+            inner class TagMatchers {
+                @Test
+                fun `should build config with path matchers`() {
+                    // Given
+                    val tagMatcher1 = TagMatcher(Tag("MYTAG1"), listOf("pathMatcher1", "pathMatcher2"))
+                    val tagMatcher2 = TagMatcher(Tag("MYTAG2"), listOf("pathMatcher3", "pathMatcher4"))
+
+                    val argv = updateMongoDbMinimumArguments().plus(
+                        listOf(
+                            "--tag", tagMatcher1.toArgument(), "--tag", tagMatcher2.toArgument()
+                        )
+                    )
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper.datamaintainConfig!!.tagsMatchers)
+                        .containsExactlyInAnyOrder(tagMatcher1, tagMatcher2)
+                }
+
+                private fun TagMatcher.toArgument(): String = "${this.tag.name}=[${this.globPaths.joinToString(", ")}]"
+            }
+
+            @Nested
+            inner class TagsList {
+                @Test
+                fun `should build config with blacklisted tags`() {
+                    testBuildConfigWithTagsList("--blacklisted-tags", DatamaintainConfig::blacklistedTags)
+                }
+
+                @Test
+                fun `should build config with whitelisted tags`() {
+                    testBuildConfigWithTagsList("--whitelisted-tags", DatamaintainConfig::whitelistedTags)
+                }
+
+                @Test
+                fun `should build config with tags to play again`() {
+                    testBuildConfigWithTagsList("--tags-to-play-again", DatamaintainConfig::tagsToPlayAgain)
+                }
+
+                private fun testBuildConfigWithTagsList(key: String, getter: KProperty1<DatamaintainConfig, Set<Tag>>) {
+                    // Given
+                    val tagsList = setOf("MYTAG", "MYOTHERTAG")
+
+                    val argv = updateMongoDbMinimumArguments().plus(
+                        listOf(
+                            key, tagsList.joinToString(",")
+                        )
+                    )
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(getter.get(configWrapper.datamaintainConfig!!))
+                        .map { it.name }
+                        .containsExactlyInAnyOrder(tagsList)
+                }
+            }
+
+            @Nested
+            inner class Rules {
+                @Test
+                fun `should build config with one rule`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus(
+                        listOf(
+                            "--rule", SameScriptsAsExecutedCheck.NAME
+                        )
+                    )
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper) {
+                        get { datamaintainConfig }.isNotNull()
+                    }
+                    expectThat(configWrapper.datamaintainConfig!!.checkRules.toList()) {
+                        hasSize(1)
+                        first().isEqualTo(SameScriptsAsExecutedCheck.NAME)
+                    }
+                }
+
+                @Test
+                fun `should build config with 2 rules`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus(
+                        listOf(
+                            "--rule", SameScriptsAsExecutedCheck.NAME,
+                            "--rule", SameScriptsAsExecutedCheck.NAME
+                        )
+                    )
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper) {
+                        get { datamaintainConfig }.isNotNull()
+                    }
+                    expectThat(configWrapper.datamaintainConfig!!.checkRules.toList()) {
+                        hasSize(2)
+                        first().isEqualTo(SameScriptsAsExecutedCheck.NAME)
+                        last().isEqualTo(SameScriptsAsExecutedCheck.NAME)
+                    }
+                }
+            }
+
+            @Nested
+            inner class AllowAutoOverride {
+                @Test
+                fun `should build default config without auto override`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments()
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper) {
+                        get { datamaintainConfig }.isNotNull()
+                    }
+                    expectThat(configWrapper.datamaintainConfig!!.overrideExecutedScripts) {
+                        isFalse()
+                    }
+                }
+
+                @Test
+                fun `should build config with auto override`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus("--allow-auto-override")
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat(configWrapper) {
+                        get { datamaintainConfig }.isNotNull()
+                    }
+                    expectThat(configWrapper.datamaintainConfig!!.overrideExecutedScripts) {
+                        isTrue()
+                    }
+                }
+            }
+
+            @ParameterizedTest
+            @EnumSource(ScriptAction::class)
+            @DisplayName("Should build config with default script action {0}")
+            fun `should build config with script action`(scriptAction: ScriptAction) {
+                // Given
+                val argv = updateMongoDbMinimumArguments().plus(
+                    listOf(
+                        "--action", scriptAction.name
+                    )
+                )
+
+                // When
+                runApp(argv)
+
+                // Then
+                expectThat(configWrapper.datamaintainConfig!!.defaultScriptAction).isEqualTo(scriptAction)
+            }
+        }
+
+        @Nested
+        inner class MongoDriverConfiguration {
+            @Nested
+            inner class MongoSaveOutput {
+                @Test
+                fun `should build config with mongo save output set to true`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus("--mongo-save-output")
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat((configWrapper.datamaintainConfig!!.driverConfig) as MongoDriverConfig)
+                        .get { saveOutput }
+                        .isTrue()
+                }
+
+                @Test
+                fun `should build config with mongo save output set to false`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments()
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat((configWrapper.datamaintainConfig!!.driverConfig) as MongoDriverConfig)
+                        .get { saveOutput }
+                        .isFalse()
+                }
+            }
+
+            @Nested
+            inner class MongoPrintOutput {
+                @Test
+                fun `should build config with create tags from folder set to true`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments().plus("--mongo-print-output")
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat((configWrapper.datamaintainConfig!!.driverConfig) as MongoDriverConfig)
+                        .get { printOutput }
+                        .isTrue()
+                }
+
+                @Test
+                fun `should build config with create tags from folder set to false`() {
+                    // Given
+                    val argv = updateMongoDbMinimumArguments()
+
+                    // When
+                    runApp(argv)
+
+                    // Then
+                    expectThat((configWrapper.datamaintainConfig!!.driverConfig) as MongoDriverConfig)
+                        .get { printOutput }
+                        .isFalse()
+                }
+            }
+        }
+    }
+
+    private fun updateMongoDbMinimumArguments(): List<String> {
+        return listOf(
+            "--mongo-uri", "mongo-uri",
+            "update-db"
+        )
+    }
+}
